Moving generic factory to it's own namespace
